+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Scopes
    - Server Hardware Type
    - Storage Pool
+   - Storage System
    - Switches
 
 #### Breaking changes